LayoutTests/media/modern-media-controls/media-documents/click-on-video-should-not-pause-expected.txt

Testing that clicking on a video in a media document does not pause it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS video.paused is false

Clicking on video
PASS video.paused is false

PASS successfullyParsed is true

TEST COMPLETE
